* refactoring * basic until_asserted functionality
