+++ tests/test_wait.py
@@ -9,6 +9,7 @@ from busypie import (FIVE_HUNDRED_MILLISECONDS, MILLISECOND,
                      set_default_timeout, wait)
 
 from backports.asyncio import run
+from tests.sleeper import assert_done_after
 
 
 def test_wait_until_condition_passed():
@@ -20,14 +21,6 @@ def test_wait_until_condition_passed_after_using_event_loop():
     run(asyncio.sleep(0.1))
     with assert_done_after(seconds=0.3) as c:
         wait().until(lambda: c.done)
-
-
-@contextmanager
-def assert_done_after(seconds):
-    sleeper = Sleeper()
-    sleeper.sleep_for(seconds)
-    yield sleeper
-    assert sleeper.done
 
 
 @pytest.mark.timeout(1)
@@ -43,20 +36,6 @@ def test_wait_until_condition_fail():
         wait().during(lambda: not c.done)
 
 
-@pytest.mark.timeout(0.6)
-def test_set_default_timeout():
-    set_default_timeout(500, MILLISECOND)
-    with pytest.raises(ConditionTimeoutError):
-        wait().until(lambda: 1 == 2)
-
-
-def test_reset_default_timeout():
-    set_default_timeout(200, MILLISECOND)
-    reset_defaults()
-    with assert_done_after(seconds=0.4) as c:
-        wait().until(lambda: c.done)
-
-
 def test_retrieve_condition_result():
     assert wait().until(lambda: 3) == 3
 
@@ -66,14 +45,3 @@ def test_nested_waits():
         return wait().until(lambda: True)
 
     assert wait().until(condition)
-
-
-class Sleeper:
-    done = False
-
-    def sleep_for(self, start):
-        Thread(target=self._wake_up_after, args=(start,)).start()
-
-    def _wake_up_after(self, seconds):
-        sleep(seconds)
-        self.done = True
